Flightsuit Movement Adjustment one
So flightsuits got the crazy TG movement speed apparently. Not sure if adjusting the Max Momentum Var will slow it down, but considering the boost speed var is setup as an acceleration it seems this would make sense.

Since we are on half of TGs movement speed I cut it in half for now. I am tired and will do more testing later on on this particular suit.