Flightsuit Movement Adjustment one (#202)
* Flightsuit Movement Adjustment one

So flightsuits got the crazy TG movement speed apparently. Not sure if adjusting the Max Momentum Var will slow it down, but considering the boost speed var is setup as an acceleration it seems this would make sense.

Since we are on half of TGs movement speed I cut it in half for now. I am tired and will do more testing later on on this particular suit.

* Effectivly halfed acceleration/momentum gain and impact lost

Since we halfed the Max Momentum from the booster mode we should half the momentum gain and loss aswell so it doesnt feel off when hitting something.