Advanced E-Gun has a slight malfunction chance again to make it less over the top.
Since nerfing the firerate would nerf the E-Guns firerate aswell and just increasing the recharge time would basicly just changing one number without any other interactions other "Duh carry another" I propose to reactivate the irradiation and EMP fail chance on the E-Gun. What it is supposed to do is to have fail ticks. Each time you use the gun or when it recharges itself (or changes its state in any way, yes this should include EMPs) it has a chance to run into a fail tick. If a fail tick happens a few things can happen. It can irradiate the user, requiring the user to get some Potassium Iodide. It can strongly irradiate the user which might result in more severe things than just a few radiation burns or it can outright cause an EMP draining its charge. The emp is, from what I have gathered not that big. So future IPCs should refrain from using that gun and pretty much ANYONE using any sort of implant since they are affected by EMPs. The reason for this nerf is that the AEG currently has an insane firerate, fast recharge and a cheap price which basicly outclasses everything sec can get short of ballistics for the occasional esword using prick. Sec will visit the chemist in medbay and ask for Anti-Rad pills containing charcoal, saline glucose and of course Potassium Iodide to heal the burns. Or you know... actually wear a damn hardsuit or heavier armor from the armory if they intend on using this weapon mainly. It is an experimental military grade weapon and should not be used without any good reason. That should stop Captains like that fucking Thali on getting one every other round because it just outclasses the E-Gun and even the Antique Laser by far.