Captains Rapier isnt a Vibroblade of death anymore.
Exactly as it says on the tin. The captains Rapier isnt a fucking Vibroblade anymore capable of dismembering crew with is because Dismemberment chance is weird as fuck and defined by IS SHARP and Armor pierce. If is Sharp is removed it shouldnt dismember organic crewmembers (well except they are skeletons for some reason. Or IPCs.). And with shouldnt I mean it doesnt.