Mind Batterer. The Epsilon Armies ultimate none lethal crowd control solution.
Well its in the code and TG removed it because of it being "never used". I used it plenty on para, its really helpfull if you have to abduct a person and torture them and dont want anyone interferring with your abduction. Also usefull for when you are on the run and sec is after you. The stun is slightly shorter than a tazer stun. Does not cause any damage its just 2 use none renewable stun. I priced this thing at 6 TC, maybe raise it to 7 considering how potent it can be if used right (or next to useless, since everyone got a 50% chance to resist it). Surplus is set to 50 so you dont wind up with 6 of those things, yet occasionally can still get it.