Lowered the Stun Chance/Knockdown Chance
Knockdown of 100 is about as potent as a stun from a stunprod.
Knockdown of 300 is twice as long as a stun from a Stunbaton.

Anything inbetween is possible if I get this right. 100 is minimum, 250 is maximum. Obviously for balancing reasons. Stun of 400 would translate into a 20 second stun. Stun of 200 into a 8-9 second stun. You see why this was kind of unhealthy before.