[WIP] Balanced Cloaking belt for Traitors
Well I have thought about other neat things traitors could use other than the standard equipment thats usually bought and I came across this "Cloaking Belt". Its a belt that makes you invisible and depletes charge in the light (which it bends arround you so you are basicly invis). 

Before the adjustment it had no effect on EMP or Ion Weapon usage. I changed that and now it should shut down if hit by any form of emp. If hit by a strong emp (direct hit of an ion weapon) it emits a flashbang like effect that stuns the belt user and anyone not wearing protective gear. Even if the belt user is flashbang proof the effect is on tile and still stuns him on the spot. Enough time for sec to take propper actions. 

Also the Itemname has changed and its flavortext made more fluffy.

The item is highly WIP and I am open for all kinds of suggestions. I actually need someone looking over this if it made any sense what I added codewise since this is actually first time adding something new to an existing item on SS13 which is not just simple changes of variables and values...

To do:
- Make it appear in the traitor uplink
- Make it play the iconic C&C Tiberian Dawn Stealth tank sound on activation (or any other easily distinguishable sound)
- Further balancing
- Fixing of bugs and errors which will inevadeably happen.