Erradicates last traces of the Barnyard Curse Spell Aviability.
Forgot with my last PR to remove the Horsemask Spell from the chance of being a one time use spellbook.
This PR fixes this. Wizards giving magic wont be able to grant this "spell" to random people on the station. This spell is even worse in the hands of none antag... it leads to alot of greytiding, killing followed by the spell user complaining in Dchat. Speaking from Para Experience.