Adjusts VTEC to lower movement, needs further balancing probably
R.speed is the var for speed modifiers apparently. This thing is reponsible why quite a few things are still running on TG movement speed.

I lowered VTECs R.speed in order to push it back in line with the regular movement speed. With this you should be faster using VTEC, just not being insanely fast. Atleast thats the theory. Cause I cannot figure for the live of it out how to take over a mob on a private server.