Fixes Issue #301. There... now you can heal dead bodies with the 3 chems you are used to
You know whats weird... everything man made can be fixed. Everything mental on a human cannot be fixed. They all stay the same.

Anyways that aside: Someone on TG was not paying attention in chemistry and said "Hey Aluminum Sulphate is a medicin, medicin doesnt work if you are dead." WRONG. It does. And now it does here. Silver Sulfadiazine has the same deal. Synthflesh is something I have gotten close to IRL with a compound... and guess what it does work regardless of tissue being alive or dead.

Now that this is aside:

#301 is fixed. Now defibs are not totally useless.