Update body.dm
At the moment, Tourette's is a crippling disability that can continuously stun you for a period of time which is obscene. The probability of being stunned paired with the duration of the stun leads to people being unable to move entirely for long periods of time, while being helplessly vulnerable. While not removing the stun, this halves the probability of it occuring and reduces the actual duration of the stun significantly. In my mind, the stun should still be far less still (Perhaps a momentary pause but with an increased frequency) but I'd not want to put forward those changes without consulting others. Thoughts?