Update book.dm
Holy Healing now plays a rarely used sound that could be vaguely taken as holy, this clears up any confusion as a result of it using the punch sound, which makes some less observant players ready to murder someone, and makes the Chaplain just slightly more holy. Curse2 is the nicest sound I could find to represent healing (Suggested by Lynxems) while Blobattack sounds wet and squelchy, which is a far nicer sound to hear when you're whacking a lifeless body with the biggest scripture that can be found on the station!

Alternatives to Blobattack would be nice if found, but again, Blobattack seems to serve it's purpose of being a wet squishy sound.