Crusher Buff
Changed base Kinetic Crusher damage to 25, was 20.
Changed Kinetic Crusher recharge to 10, was 15
Changed Crusher Trophy drop rate to 20, was 5

The increase to the base damage and charge speed on the crusher allow it to still stay very competitive when compared to its' counterparts the Adv. Plasma Cutter, and Kinetic Accelerator that pretty much every miner will always time over the Crusher currently.

The increase to Trophy drop rates change it from the very unlikely 1 in 20 mobs being killed for a single trophy, to 1 in 5 mobs on average. I have yet personally to ever see a crusher trophy be dropped and neither has anyone else from non trivial monsters. Trivial monsters such as bubblegum remain the same with requiring 60% of damage dealt to them to have been from a kinetic crusher.