Clockwork Spear + (Un)Holy Fire Weapons
This will reduce the pitchfork's brute damage by 8 but now the weapon deals 5 burning damage as a trade-off. I believe that the trade is worth it, but would would be averse to increasing the fire damage to 8 if it turns out that the null rod becomes too weak as a result. This seems unlikely to me. The Pitchfork will no longer fit in the bag of the Chaplain, or any thieves.  

Also adds a reskin of the Pitchfork, a Clockwork Spear, which has it's own sound effects and list of attack verbs. Clockwork cultists will see this weapon and the base null rod differently to normal crew now, having a clock-cult description of it instead of the normal one.