The power of [Deity_Name] Compels you!
Adds more phrases for the Chaplain to shout when removing runes, as storming into a base and spamming the same stuff will be spammy af.

Purely fluff.

I’m considering adding a colour or effect to it if possible. The idea of Holytongue could work too, but I’d rather make sure that this gets through first at all.